Time: 1 ms (97.92%) | Memory: 42.6 MB (60.20%) - LeetSync

+++ 69-sqrtx/sqrtx.java
@@ -1,9 +1,6 @@
 class Solution {
     public int mySqrt(int x) {
-
-        int s = (int) Math.sqrt(x);
-
-        return s;
-        
+        double s = Math.sqrt(x);
+        return (int) (s);
     }
 }
